refactor: move prevInstruction and nextInstruction onto the disassembler
Neither depends on the debugger UI: one is the disassembler's own step
length, the other a scoring heuristic over its output, so they live on
Disassemble6502 with pc as an explicit parameter and the Debugger
methods delegate. Their tests move to the opcodes' own test file.
