+++ src/6502.opcodes.js
@@ -1058,6 +1058,52 @@ class Disassemble6502 {
         }
         return [opcode, addr + 1];
     }
+
+    nextInstruction(address) {
+        return this.disassemble(address)[1] & 0xffff;
+    }
+
+    /**
+     * Guesses the address of the instruction before the given one by scoring
+     * each possible run of instructions leading up to it: a point per "common"
+     * instruction (loads, stores, branches, compares, arithmetic and
+     * carry-set/clear that avoid unusual indexing modes) and a large boost for
+     * a run that passes through pc. The highest-scoring run wins and its last
+     * instruction is the answer.
+     * Good test cases:
+     *   Repton 2 @ 2cbb
+     *   MOS @ cfc8
+     * also, just starting from the back of ROM and going up...
+     */
+    prevInstruction(address, pc) {
+        const commonInstructions =
+            /(RTS|B..|JMP|JSR|LD[AXY]|ST[AXY]|TA[XY]|T[XY]A|AD[DC]|SUB|SBC|CLC|SEC|CMP|EOR|ORR|AND|INC|DEC).*/;
+        const uncommonInstrucions = /.*,\s*([XY]|X\))$/;
+
+        address &= 0xffff;
+        let bestAddr = address - 1;
+        let bestScore = 0;
+        for (let startingPoint = address - 20; startingPoint !== address; startingPoint++) {
+            let score = 0;
+            let addr = startingPoint & 0xffff;
+            while (addr < address) {
+                const result = this.disassemble(addr);
+                if (addr === pc) score += 10;
+                if (result[0].match(commonInstructions) && !result[0].match(uncommonInstrucions)) {
+                    score++;
+                }
+                if (result[1] === address) {
+                    if (score > bestScore) {
+                        bestScore = score;
+                        bestAddr = addr;
+                        break;
+                    }
+                }
+                addr = result[1];
+            }
+        }
+        return bestAddr;
+    }
 }
 
 function makeCpuFunctions(cpu, opcodes, is65c12, cycleAccurate = true) {
