+++ tests/unit/test-web-debug.js
@@ -1,38 +1,10 @@
 // @vitest-environment jsdom
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { Debugger, prevInstruction } from "../../src/web/debug.js";
-import { Cpu6502 as cpu6502Opcodes } from "../../src/6502.opcodes.js";
+import { Debugger } from "../../src/web/debug.js";
 import { fake6502 } from "../../src/fake6502.js";
 import { FakeVideo } from "../../src/video.js";
 import { domFromIndexHtml, teardownDom } from "./helpers.js";
-
-describe("prevInstruction", () => {
-    const mem = new Uint8Array(0x10000);
-    const disassembler = cpu6502Opcodes({ peekmem: (addr) => mem[addr & 0xffff] }).disassembler;
-    const disassemble = (addr) => disassembler.disassemble(addr);
-
-    beforeEach(() => {
-        mem.fill(0xea);
-    });
-
-    it("returns the previous instruction in an unambiguous run", () => {
-        expect(prevInstruction(disassemble, 0x1000, 0x2002)).toBe(0x2001);
-    });
-
-    it("prefers the run of instructions that passes through the program counter", () => {
-        // Five LDA $41 pairs then $ad give a misaligned parse whose LDA $41a9 swallows
-        // the real LDA #$41 at 2000; only the boost for a run through pc outscores it.
-        mem[0x2000] = 0xa9;
-        mem[0x2001] = 0x41;
-        for (let i = 0; i < 5; i++) {
-            mem[0x1ff5 + i * 2] = 0xa5;
-            mem[0x1ff6 + i * 2] = 0x41;
-        }
-        mem[0x1fff] = 0xad;
-        expect(prevInstruction(disassemble, 0x2000, 0x2002)).toBe(0x2000);
-    });
-});
 
 describe("Debugger", () => {
     let cpu;
